refactor: move history orchestration to ui kit
