+++ packages/prices-api/src/llamalend/index.ts
@@ -1,5 +1,5 @@
 import { addQueryString, fetchJson as fetch } from '@primitives/fetch.utils'
-import { getHost, type Chain, type Options, type SnapshotQueryParams } from '..'
+import { getHost, type Chain, type Options } from '..'
 import { type Endpoint } from '../llamma'
 import * as Schema from './schema'
 export { tvl } from './util'
@@ -42,7 +42,13 @@ export async function getMarkets(
 export async function getSnapshots(
   chain: Chain,
   marketController: string,
-  params: SnapshotQueryParams = { fetch_on_chain: true, agg: 'day', limit: 100 },
+  params: {
+    agg?: string
+    fetch_on_chain?: boolean
+    limit?: number
+    start?: number
+    end?: number
+  } = { fetch_on_chain: true, agg: 'day', limit: 100 },
   options?: Options,
 ) {
   const host = getHost(options)
