+++ packages/prices-api/src/usd-price/index.ts
@@ -1,13 +1,12 @@
 import type { Address } from '@primitives/address.utils'
 import { addQueryString, fetchJson as fetch } from '@primitives/fetch.utils'
 import { getHost, type Chain, type Options } from '..'
-import { fetchChunkedTimeSeries, getTimeRange } from '../timestamp'
+import { getTimeRange } from '../timestamp'
 import * as Schema from './schema'
 
 export type * from './schema'
 
-// The endpoint returns at most 300 rows. A 299-day inclusive range contains at most 300 daily buckets.
-const MAX_HISTORY_RANGE_DAYS = 299
+type GetUsdPriceHistoryRangeParams = { start: number; end: number }
 
 /** Fetch USD price for a token. */
 export async function getUsdPrice(blockchainId: Chain, contractAddress: Address, options?: Options) {
@@ -24,18 +23,21 @@ export async function getUsdPriceHistory(
   days: number,
   options?: Options,
 ) {
-  const host = getHost(options)
   const range = getTimeRange({ daysRange: days })
+  return getUsdPriceHistoryRange(blockchainId, contractAddress, range, options)
+}
 
-  return fetchChunkedTimeSeries({
-    range,
-    maxDays: MAX_HISTORY_RANGE_DAYS,
-    order: 'asc',
-    fetchChunk: async range => {
-      const params = { interval: 'day', ...range }
-      const url = `${host}/v1/usd_price/${blockchainId}/${contractAddress}/history${addQueryString(params)}`
-      const response = await fetch(url)
-      return Schema.getUsdPriceHistoryResponse.parse(response)
-    },
-  })
+/** Fetch daily USD prices for one bounded time range. */
+export async function getUsdPriceHistoryRange(
+  blockchainId: Chain,
+  contractAddress: Address,
+  range: GetUsdPriceHistoryRangeParams,
+  options?: Options,
+) {
+  const host = getHost(options)
+  const params = { interval: 'day', ...range }
+  const url = `${host}/v1/usd_price/${blockchainId}/${contractAddress}/history${addQueryString(params)}`
+  const response = await fetch(url)
+
+  return Schema.getUsdPriceHistoryResponse.parse(response)
 }
